% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-scores.R
\name{parse_scores}
\alias{parse_scores}
\title{Parses the scores of the html wiki season page.}
\usage{
parse_scores(html, season)
}
\description{
\code{parse_scores}
}
\details{
The xpath works like this:

1) Look for a Results id.

2) Go to the parent. I'm not specific here about what the parent should be like because it can be h2, h3, ...

3) Look for the next following sibling. Next because [1] is specified.

4) Get each cell of the table.
}
